TRAC#7497 Adds pages list
