+++ app/comicsite/templates/admin/comicmodels/page/change_form.html
@@ -6,7 +6,7 @@
         <div class="submit-row">
             <p class="deletelink-box"></p>
             <a class="editPageLink"
-               href="{% url 'pages:page-detail' original.comicsite.short_name original.title %}">Cancel and return to
+               href="{% url 'pages:detail' original.comicsite.short_name original.title %}">Cancel and return to
                 page</a>
             <input type="submit" value="Save and return to page" name="save_goto_page">
