+++ app/comicsite/templatetags/comic_templatetags.py
@@ -867,8 +867,8 @@ class InsertFileNode(template.Node):
 
         # any relative link inside included file has to be replaced to make it work within the COMIC
         # context.
-        base_url = reverse('pages:insertpage-detail', kwargs={
-            'site_short_name': currentpage.comicsite.short_name,
+        base_url = reverse('pages:insert-detail', kwargs={
+            'challenge_short_name': currentpage.comicsite.short_name,
             'page_title': currentpage.title,
             'dropboxpath': "remove"})
         # for some reason reverse matching does not work for emtpy dropboxpath (maybe views.dropboxpage
@@ -1022,8 +1022,8 @@ class InsertGraphNode(template.Node):
 
         # any relative link inside included file has to be replaced to make it work within the COMIC
         # context.
-        base_url = reverse('pages:insertpage-detail', kwargs={
-            'site_short_name': context.page.comicsite.short_name,
+        base_url = reverse('pages:insert-detail', kwargs={
+            'challenge_short_name': context.page.comicsite.short_name,
             'page_title': context.page.title,
             'dropboxpath': "remove"})
         # for some reason reverse matching does not work for emtpy dropboxpath (maybe views.dropboxpage
