+++ app/pages/templates/comicmodels/page_list.html
@@ -1,0 +1,16 @@
+{% extends "site.html" %}
+
+{% block pagecontent %}
+
+    <h2>Pages for {{ site.short_name }}</h2>
+
+    <ul>
+
+    {% for page in object_list %}
+        <li>
+            {{ page.title }}
+        </li>
+    {% endfor %}
+
+    </ul>
+{% endblock %}
