+++ app/tests/utils.py
@@ -1,10 +1,14 @@
+from typing import Callable
 from urllib.parse import urlparse
 
 from django.conf import settings
-from django.test import RequestFactory
+from django.contrib.auth.models import AnonymousUser
+from django.test import RequestFactory, Client
 from django.views.generic import View
 
 from comicmodels.models import ComicSite
+from comicsite.core.urlresolvers import reverse
+from tests.factories import SUPER_SECURE_TEST_PASSWORD
 
 
 def assert_redirect(uri: str, *args):
@@ -30,3 +34,165 @@ def assert_status(code: int,
 
     assert response.status_code == code
     return request, response
+
+
+def get_view_for_user(*,
+                      viewname: str = None,
+                      challenge: ComicSite = None,
+                      client: Client,
+                      method: Callable,
+                      pk: str = None,
+                      user: settings.AUTH_USER_MODEL = None,
+                      url: str = None,
+                      **kwargs):
+    """ Returns the view for a particular user """
+    if url is None:
+
+        reverse_kwargs = {'challenge_short_name': challenge.short_name}
+
+        if pk is not None:
+            reverse_kwargs['pk'] = pk
+
+        url = reverse(viewname, kwargs=reverse_kwargs)
+
+    elif viewname:
+        raise AttributeError('You defined both a viewname and a url, only '
+                             'use one!')
+
+    if user and not isinstance(user, AnonymousUser):
+        client.login(username=user.username,
+                     password=SUPER_SECURE_TEST_PASSWORD)
+
+    response = method(url, **kwargs)
+
+    if user:
+        client.logout()
+
+    return response
+
+
+def assert_viewname_status(*, code: int, **kwargs):
+    """ Asserts that a viewname for challenge_short_name and pk returns status
+    code `code` for a particular user """
+
+    response = get_view_for_user(**kwargs)
+
+    assert response.status_code == code
+    return response
+
+
+def assert_viewname_redirect(*,
+                             url: str,
+                             **kwargs):
+    """ Asserts that a view redirects to the given url. See
+    assert_viewname_status for kwargs details """
+    response = assert_viewname_status(code=302, **kwargs)
+    redirect_url = list(urlparse(response.url))[2]
+    assert url == redirect_url
+    return response
+
+
+def validate_admin_only_view(*,
+                             two_challenge_set,
+                             client: Client,
+                             **kwargs):
+    """ Assert that a view is only accessible to administrators for that
+    particular challenge """
+    # No user
+    assert_viewname_redirect(
+        url=settings.LOGIN_URL,
+        challenge=two_challenge_set.ChallengeSet1.challenge,
+        client=client,
+        method=client.get,
+        **kwargs
+    )
+
+    tests = [
+        (403, two_challenge_set.ChallengeSet1.non_participant),
+        (403, two_challenge_set.ChallengeSet1.participant),
+        (403, two_challenge_set.ChallengeSet1.participant1),
+        (200, two_challenge_set.ChallengeSet1.creator),
+        (200, two_challenge_set.ChallengeSet1.admin),
+        (403, two_challenge_set.ChallengeSet2.non_participant),
+        (403, two_challenge_set.ChallengeSet2.participant),
+        (403, two_challenge_set.ChallengeSet2.participant1),
+        (403, two_challenge_set.ChallengeSet2.creator),
+        (403, two_challenge_set.ChallengeSet2.admin),
+        (200, two_challenge_set.admin12),
+        (403, two_challenge_set.participant12),
+        (200, two_challenge_set.admin1participant2),
+    ]
+
+    for test in tests:
+        assert_viewname_status(
+            code=test[0],
+            challenge=two_challenge_set.ChallengeSet1.challenge,
+            method=client.get,
+            client=client,
+            user=test[1],
+            **kwargs
+        )
+
+
+def validate_admin_or_participant_view(*,
+                                       two_challenge_set,
+                                       client: Client,
+                                       **kwargs):
+    """ Assert that a view is only accessible to administrators or participants
+    of that particular challenge """
+    # No user
+    assert_viewname_redirect(
+        url=settings.LOGIN_URL,
+        challenge=two_challenge_set.ChallengeSet1.challenge,
+        client=client,
+        method=client.get,
+        **kwargs
+    )
+
+    tests = [
+        (403, two_challenge_set.ChallengeSet1.non_participant),
+        (200, two_challenge_set.ChallengeSet1.participant),
+        (200, two_challenge_set.ChallengeSet1.participant1),
+        (200, two_challenge_set.ChallengeSet1.creator),
+        (200, two_challenge_set.ChallengeSet1.admin),
+        (403, two_challenge_set.ChallengeSet2.non_participant),
+        (403, two_challenge_set.ChallengeSet2.participant),
+        (403, two_challenge_set.ChallengeSet2.participant1),
+        (403, two_challenge_set.ChallengeSet2.creator),
+        (403, two_challenge_set.ChallengeSet2.admin),
+        (200, two_challenge_set.admin12),
+        (200, two_challenge_set.participant12),
+        (200, two_challenge_set.admin1participant2),
+    ]
+
+    for test in tests:
+        assert_viewname_status(
+            code=test[0],
+            challenge=two_challenge_set.ChallengeSet1.challenge,
+            method=client.get,
+            client=client,
+            user=test[1],
+            **kwargs
+        )
+
+
+def validate_open_view(*,
+                       challenge_set,
+                       client: Client,
+                       **kwargs):
+    tests = [
+        (200, None),
+        (200, challenge_set.non_participant),
+        (200, challenge_set.participant),
+        (200, challenge_set.participant1),
+        (200, challenge_set.creator),
+        (200, challenge_set.admin)
+    ]
+
+    for test in tests:
+        assert_viewname_status(code=test[0],
+                               challenge=challenge_set.challenge,
+                               method=client.get,
+                               client=client,
+                               user=test[1],
+                               **kwargs)
